safari-600.7-branch/LayoutTests/fast/dom/Geolocation/watchPosition-unique-expected.txt

Tests that navigator.geolocation.watchPosition returns unique results within its script execution context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS watchID1 > 0 is true
PASS watchID2 > 0 is true
PASS watchID3 > 0 is true
PASS watchID1 !== watchID2 is true
PASS watchID1 !== watchID3 is true
PASS watchID2 !== watchID3 is true
PASS successfullyParsed is true

TEST COMPLETE
